Code Cleanup.
